Add back the classes as part of HTML post-render-html-fixups
This makes more sense to get back the behavior from previous Pandoc.
