+++ src/resources/filters/crossref/crossref.lua
@@ -61,7 +61,6 @@ import("../quarto-init/resourcerefs.lua")
 
 import("../normalize/flags.lua")
 import("../normalize/normalize.lua")
-import("../normalize/pandoc-tables.lua")
 import("../normalize/parsehtml.lua")
 import("../normalize/extractquartodom.lua")
 import("../normalize/astpipeline.lua")
