+++ src/resources/filters/normalize/astpipeline.lua
@@ -22,9 +22,6 @@ function quarto_ast_pipeline()
     }
   end
   return {
-    -- identify pandoc table
-    { name = "normalize-pandoc-table", filter = pandoc_tables() },
-
     { name = "normalize-table-merge-raw-html", filter = table_merge_raw_html() },
 
     -- this filter can't be combined with others because it's top-down processing.
